views: replace continue_in_background decorator with explicit daemonize call
Previously views got daemonized directly on startup. That meant that a view
which got started accidentally and waits for an input event for example, would
run til the server shuts down.
Since the resources get limited by the scheduler, this can lock the server
pretty quickly.
Now it is possible to daemonize not until the view actual does something that
should run in background.

Signed-off-by: Florian Scherf <f.scherf@pengutronix.de>