input events: always send a full node description
Previously sending an input event, containing a node id for an node that is
already deleted or overridden on the server side, would result in the input
event class raising a ValueError. With the new API this can be intended, so
the client sends now a full node description on every input event fired and the
server tries to find the corresponding node. In the worst case the node is None
but no error is raised.

Signed-off-by: Florian Scherf <f.scherf@pengutronix.de>