run_server: don't change directory when starting server
Previously run_server would take project_root and change the global working
directory to this location. This causes problems when Lona gets used with
pytest.

This patch adds path and importing infrastructure to the server class and
patches all code that depend on the old importing infrastructure to make
relative paths and imports possible.

Signed-off-by: Florian Scherf <f.scherf@pengutronix.de>