protocol: add prefix to all Lona messages
Previously the metric to decide if an websocket message is a Lona message was:

  1. It has to be JSON
  2. It has to JSON decode to a list
  3. The first list item has to be an integer

Since Lona allows websocket middlewares to have support for all sorts of
messages, false positives of this metric were a problem.

This patch adds the prefix "lona:" to all messages, to make recognition easier
and more robust.

Signed-off-by: Florian Scherf <f.scherf@pengutronix.de>