html: data binding: fix edit war bug
When using data binding the data binding widgets catch their own change events
and synchronize there internal value. When setting their value, a HTML patch,
containing the set value, gets created that gets distributed to all connections
that are connected to the runtime. By default the input timeout is set to
300ms. When the user is typing fast, it can happen that a patch gets applied
when the user is still typing, resetting the input to a previous state.
For the user it looks like the input looses characters while typing.

The solution for this problem is to don't send patches back to users who
issued them.

Signed-off-by: Florian Scherf <f.scherf@pengutronix.de>