html: data binding: fix TextInput.handle_change
TextInput.handle_change gets called at the correct time, but the value of
TextInput.value gets not set before, which is counter-intuitive.

This patch sets TextInput.value in TextInput.handle_input_event() and then
calls TextInput.handle_change(). This patch also adds a test view for this
feature to the test project.

Signed-off-by: Florian Scherf <f.scherf@pengutronix.de>