input events: fix custom not node related input events
Frontend widgets can fire custom input events which are not related to a node.

    https://lona-web.org/end-user-documentation/html.html#firing-custom-input-events)

Previously the frontend widget sent its widget id with every input event and
when the node id was empty the backend widget got identified by it. This broke
between 1.4.1 and 1.5.

Since all nodes can handle input events now, this extra field for the widget id
is not needed anymore. This patch removes the widget id from the event payload
and lets frontend widgets send their own id if no node is related to the event.

Signed-off-by: Florian Scherf <f.scherf@pengutronix.de>