html: fix node list operations
Previously nodes had a special __eq__() method that made high level comparisons
between nodes possible.

    Div(a=1) == Div(a=1) -> True
    Div(a=1) == Div(b=1) -> False

This feature was intended to be used for testing purposes in the future.

The problem with this implementation is that methods like list.index() or
list.remove() use __eq__() to find an list item.

    d1 = Div()
    d2 = Div()
    l = [d1, d2]

    l.index(d1) -> 0
    l.index(d2) -> 0

This breaks patch generating and a lot of potential application code.

Signed-off-by: Florian Scherf <f.scherf@pengutronix.de>