feat(tests): add eventually() to wait for expected state
New function eventually() allows to wait for an expected state in tests. Iterator and context managers are used to be able repeating assertions.

Alternative approach is to accept function as an argument, and call it several times. But in this case user has to create a function inside test to make an assert. That's because assert is a statement and python doesn't allow it inside the lambda.

Signed-off-by: Marat Reymers <marat.maratori@gmail.com>