feat(protocol): setup ping messages
Modern browsers like Chrome close websockets after a preconfigured timeout of
around five minutes of inactivity to save energy. This can lead to all sorts
of bad user experience, because all important state is part of the view in
Lona.
This patch adds ping and pong messages to the protocol and client
infrastructure to send pings in a configurable interval.

Signed-off-by: Florian Scherf <f.scherf@pengutronix.de>