fix(server): fix slow down when removing connections
Previously ViewRuntimeController.remove_connection was called directly in a
coroutine, which caused slow downs and in some test cases even lockups, because
ViewRuntimeController.remove_connection pulls a HTML lock implicitly. This lock
is shared with application code, which means the coroutine has to wait for the
view to release the lock in some cases, which locks the ioloop.
This patch wraps the remove_connection call in a function and runs it in a
worker, to release the ioloop as soon as possible.

Signed-off-by: Florian Scherf <f.scherf@pengutronix.de>