feat(client): pre compiler: add better debug mode
The JavaScript client uses a custom pre compiler that implements includes but
no map files. That means it can be hard sometimes to debug the client because
the Chrome Inspector points to lines in the code that do not exist in the
source code.

This patch adds a proper debug mode that serves all files individually without
merging them together. This is simpler than implementing map files and makes
Chrome Inspector usable again.

Signed-off-by: Florian Scherf <mail@florianscherf.de>