logo: fix bleeding background
Some renderers have bugs when two areas are right next to each other,
and render each area on their own with antialiasing in relation to the
background color. This has the disadvantageous effect that the
background color bleeds through to the foreground, which in our case is
mostly visible on dark backgrounds.

To fix this, add a 1px outline with the same color as the fill color to
each of those areas, so that neighboring areas effectively overlap with
each other, thereby covering the background.