fix(views): fix page titles on daemonized views
Previously the message for changing the page title using View.show() or
View.set_title() was send only once. This is fine for non-daemon views but on
daemon views the title has to be resend everytime the user reconnects, or a
second user connects.

This patch adds a title field to the HTML document, so it gets saved beyond
reconnects and adds title resends to the view runtime.

Signed-off-by: Florian Scherf <mail@florianscherf.de>