feat(server/routing): Add support for trailing wildcards
This change adds support for wildcard routes.
Wildcards can only be used at the end of a route.

A wildcard can be used to route all URLs beginning with a
given path to a view or a Django application.

Consider the following example for the Django Admin:
The easy way to route the Django Admin is to have a fallback route at
the end of the routing table.
But this is not possible if there is another route that matches the
URL.

In this case a wildcard match can be used:

```
routes = [
    # (...)

    # More specific match
    Route('/admin(*)', wsgi_handler, http_pass_through=True),

    # Broad match
    Route('/<output>(/)', 'views/home.py::HomeView'),
]
```

Signed-off-by: Chris Fiege <chris@tinyhost.de>