doc: move documentation to prefix 1.x
Lona 2 is coming! \o/

This patch moves the current documentation to /1.x/ to make room for
documentation for Lona 2.
This is achieved by adding a Flamingo plugin that moves all existing articles,
and a redirect rule file that adds rules to redirect the old URLs to the new,
prefixed ones.

Signed-off-by: Florian Scherf <mail@florianscherf.de>