feat(LonaServer): Allow to set aiohttp client_max_size
With this change it is now possible to set the aiohttp
``client_max_size`` during the creation of the ``Application``. This
value controls the maximum size of a POST request in bytes.

This change basically moves the creation of the aiohttp ``Application``
from ``run_server.py`` and ``app.py`` into the ``LonaServer``.

Signed-off-by: Chris Fiege <chris@tinyhost.de>