client: use ES06 imports instead of Python based pre compiler
The sole reason for the client pre compiler was to add Python constants and
Javascript imports to the vanilla Javascript client implementation.

Since all major browsers support ES06 imports now, and Python constants can
also resolved in the templating stage, the client pre compiler is obsolete now.

This patch ports the client to ES06 imports and moves constant resolving to
the templating stage.

Signed-off-by: Florian Scherf <mail@florianscherf.de>