client: fix node cache problem
Previously the node cache got cleaned out after every patch that was applied.
In some cases that resulted in situations in which nodes got cleaned out of the
cache before they were applied to the Dom.
When a patch came in, for an node that was not present in the node cache, the
client crashed.

This patch removes the cache cleaning for every patch, and adds one cache
cleaning call to the end of every patch stack.

Signed-off-by: Florian Scherf <mail@florianscherf.de>