client: remove cache cleaning issue mitigations
The client had some known issues regarding node cache lookups, that couldn't be
pinned down exactly until recently.

To mitigate these issues, the client had some additional checks if a node
could be looked up or not, to not crash while rendering. Since these issues
are resolved, and we want to search for more, this patch disables all these
checks and lets the client crash if a node can't be looked up.

Signed-off-by: Florian Scherf <mail@florianscherf.de>