html: Node: add frontend widget capabilities
Currently the HTML API implements Nodes and Widgets. Nodes describe a "real"
HTML node that can be rendered in the browser, widgets describe a collections
of such nodes.
These two levels of abstraction were implemented to make high-level widgets
with more than one root node possible.

The multi root node support adds complexity to the rendering client in the
browser. It turned out that this feature may be useful in some cases,
but it is not necessary. Therefore the community decided to remove it in the
future to make the browser client more compact.

This patch adds all frontend widget capabilities of widgets to the base node
class, and a test for the new feature. This will make widgets obsolete.

Signed-off-by: Florian Scherf <mail@florianscherf.de>