html: HTML: port from widget to function
BREAKING CHANGE: This patch changes the return value and type of
                 lona.html.HTML

The community decided to merge the capabilities of nodes and widgets into one,
simpler base class. Until this point Lona came with only one widget in it's
standard library (lona.html.HTML).

To make widgets completely obsolete, this patch ports lona.html.HTML from a
widget to a function.

Signed-off-by: Florian Scherf <mail@florianscherf.de>