build system: test: move to docker
The Lona test suite uses Microsoft Playwright for its browser tests, which
actually expects to run on Ubuntu. Also the test suite has to run sudo to
install all playwright dependencies. This is fine in a Github action, but
not on a developer machine.

This patch adds all necessary docker files to run the test suite in a
Playwright docker container and wraps make target into a docker-compose
command.

Signed-off-by: Florian Scherf <mail@florianscherf.de>