test project: build system: update make targets
Previously the Makefile used normal perquisites for the Python virtualenv
directory. Make checks the modified time of this directory on every build.
On some systems this modified time changes every time the directory gets
accessed, which leads to false-positives.

This was previously accounted for by adding a helper file inside the
virtualenv. This is unnecessary because make supports order-only-perquisites,
which only ensure the virtualenv exists before running `server` for example.

This patch changes all perquisites to directories to order-only-perquisites
This patch does some cleanup too by adding a default target to `server` and
removing an out of date target named `message-broker`.

Signed-off-by: Florian Scherf <mail@florianscherf.de>