html: select: refactor API
Previously the API of `html.Select` was quite confusing because it's main
control mechanism over it's options and their values were `html.Select.value`
and `html.Select.values`. `values` was represented as a list of tuples, which
got parsed into `html.Option` objects with their values and attributes set.
Values always got converted to strings, which is the correct behavior, from a
browsers perspective, but it was surprising and inconvenient.

This patch splits the concerns of options and values into separate properties,
and uses high-level node options to represent options.

This patch also changes the clients input event code to send the index of
selected options instead of their values. The values get still string converted
to make the options look correct in the browser, but the server can use the
original, unconverted values.

Old API:

    >>> Select(
        values=[
            # value, label, is_selected
            ('foo', 'Foo', True),
            ('bar', 'Bar', False),
        ],
    )

New API:

    >>> select = Select(
    	Option('Option 1', value='1'),
    	Option('Option 2', value=2),
    	Option('Option 3', value=3.0, selected=True),
    )

    >>> select.value
    3.0
    >>> select.value = '1'  # selects the first option
    >>> select.options[0].selected = True  # selects the first option
    >>> select.options[-1].disabled = True  # disables the last option

Signed-off-by: Florian Scherf <mail@florianscherf.de>