html: remove Widget API
Previously Lona defined two types of nodes: Nodes that can be rendered in the
browser (text nodes and elements), and collections of nodes that can be
rendered (widgets). This distinction was made to make components (for example a
pop-up component) with multiple root nodes possible.

This feature is useful in some cases but brings much complexity and error
potential into the JavaScript client. Also the implications of enforcing every
component to have exactly one root node, are not big enough to justify this
level of complexity.

This patch removes the old widget rendering API on server and client entirely,
and changes the wording of the component API: Previously we had "Widget" on the
server and "Frontend Widget" which described a JavaScript class, corresponding
the widget on the server.
After this patch all server side HTML representations are called "Node" or
"Component" for nodes that represent high-level components like pop-ups, and
all client side code is called "Widget". Every node can define a widget using
`Node.WIDGET` which contains a string with the name of the JavaScript class.

Signed-off-by: Florian Scherf <mail@florianscherf.de>