pytest: debug_interactive: disable stdin and stdout capturing during runtime
`lona.pytest.LonaContext.debug_interactive` starts a rlpython shell that reads
and writes to stdin and stdout, which are captured by pytest by default.

Previously this had to be disabled by hand, by setting `-s` in the pytest
command line (or respective pytest config variable) to make the shell work.

This patch adds code that disables capturing for the runtime of
`lona.pytest.LonaContext.debug_interactive` and re-enables it afterwards.

Signed-off-by: Florian Scherf <mail@florianscherf.de>