tests: add rendering test
This patch adds a view to the test project, which uses all HTML rendering
features step-by-step, and adds a test to the test suite which runs the view
and checks if the server-side and the client-side HTML match for the duration
of the view.

This indirection over the test project is made because this way the test view
is easier to run in a development setup.

Signed-off-by: Florian Scherf <mail@florianscherf.de>