fix(collect-static): remove usage of old client pre compiler API
This patch fixes a regression, introduced in
7d8b5513b72d ("client: remove client pre compiler").

Before 7d8b5513b72d, the collect-static mechanism used the client pre-compiler
to get a list of all client JavaScript files to copy to the output. This API
is no longer available, because the client pre-compiler was removed.

This patch creates a new method in the static file loader, which returns all
found static files, and changes collect-static to use it.

Signed-off-by: Florian Scherf <mail@florianscherf.de>