client: dom updater: fix HTML id remove method
Previously LonaDomUpdater.remove_id used Array.pop to remove a specific HTML
id, but that is not was Array.pop does. Array.pop removes the last item of an
Array, which resulted in the wrong HTML id got removed.

This patch replaces the Array.pop call with a simple filter.

Signed-off-by: Florian Scherf <mail@florianscherf.de>