client2: merge LonaDomRenderer, LonaDomUpdater and LonaWidgetDataUpdater
Previously the rendering code was scattered over multiple classes and source
files to increase readability and add some separation of concerns.
For the most part this worked, but it lead to some spaghetti code.

Since support for the old Widget HTML API was removed, the rendering code is
much smaller and easier to read anyway.

This patch merges all code, concerned with rendering, into one module and
one class named "LonaRenderingEngine".

Signed-off-by: Florian Scherf <mail@florianscherf.de>