client2: rendering: add safety check to node cache
Previously the rendering code sometimes accessed the node cache directly,
instead of using `_get_node()`. JavaScript (being JavaScript) returns
`undefined` if no node with the given node id exists.
So looking up an unknown node id "worked" but the code then crashed when trying
to patch the retrieved node, which was hard to debug.

This patch adds a node id check to `_get_node()`, which throws an exception,
when an unknown node id was given.

Signed-off-by: Florian Scherf <mail@florianscherf.de>