client: rendering: html updates: remove obsolete node cache cleaning calls
Previously the rendering engine cleaned the node cache after every applied
patch. That made sense back then, because when the rendering still supported
widgets, node cache lookups got more expensive with every cached node.

Since the widgets are removed, and the node cache is a simple key-value-store,
this is obsolete.

Signed-off-by: Florian Scherf <mail@florianscherf.de>