client2: rendering: id reset: remove obsolete loop
When the server wants to reset an id attribute, it sends a list of ids like

    ['foo', 'bar', 'baz']

Previously the reset code iterated over the list of ids, but passed the whole
list into `_add_id()`, every iteration.
That worked because `_add_id()` uses `Array.concat()`, which accepts scalar
types and arrays. When an Array is given, every item in the array is added to
the original array.

This patch changes the code to iterate over the list and call `_add_id()`
for every id.

Signed-off-by: Florian Scherf <mail@florianscherf.de>