client: rendering: use proper types and methods for node cache cleaning
Previously the rendering engine used objects instead of maps for the node
cache, which can't be cleared easily. To get around this, the old code just
overwrote the variables that hold the node cache, to ensure the cache
is empty.

This is dangerous, because someone could hold a reference to the old variables.

This patch replaces all objects in the node cache with maps, uses `Map.clear()`
to clear them and replaces the overriding of arrays with `Array.length = 0`.

Signed-off-by: Florian Scherf <mail@florianscherf.de>