input events: add support for target node info
In JavaScript, when an event listener for a click event, is attached to a
node, the resulting event can be originally issued by one of its child nodes,
but catched by the node that defined the event listener.

Previously, when setting up events on a Lona node, there was no way to
determine if an event was issued by this exact node or by one of its child
nodes.

This resulted in problems when using clickable nodes inside clickable nodes,
which can be a valid use-case, for example for clickable backdrops that contain
buttons.

This patch adds a new property to input events called `target_node` which is
the equivalent to `event.target` in JavaScript.

Signed-off-by: Florian Scherf <mail@florianscherf.de>