html: AbstractNode: fix unmounting and loop detection issues
Lona nodes have to be unique, because they are meant to represent exactly one
node in the browser DOM. This means, when a node gets mounted into a node tree,
it has to be unmounted at its previous parent node tree, if present.

Previously this mechanism was flawed, and there were scenarios in which a node
could appear in multiple node trees, or appear multiple times in the same node
tree. In these cases the loop detection sometimes ended up in an endless loop.

Example 1:

    >> from lona.html import Div
    >> div1 = Div()
    >> div2 = Div()
    >> div1.append(div2)
    >> div1.append(div2)
    >> div1.append(div2)
    >> len(div1.nodes)  # should be 1 but it is 3

Example 2:

    >> from lona.html import Div
    >> div = Div(Div(Div(Div())))
    >> div[0][0][0].append(div[0][0])  # hangs in an endless loop

This patch fixes these issues by fixing the loop detection, fixing the parent
checks in `NodeList._prepare_node` and adding some tests for these kinds of
issues.

Signed-off-by: Florian Scherf <mail@florianscherf.de>