views: add daemonizing support for short running views
Previously daemonizing required views their `handle_request()` method to run as
long as they wanted to be daemonized.

Initially daemonizing was planned as a feature for views that run long running
operations and push a progress in the process. Daemonizing short running views
only worked when a busy loop or a long sleep or was inserted, to artificially
prevent `handle_request()` from finishing.

This patch changes the view runtime checks so that daemonized views can be
finished without getting removed from the server.

Previously a view got daemonized by calling `LonaView.daemonize()` and
"undaemonized" and removed from the server by simply returning from
`handle_request()`. This patch adds a new boolean property called
`LonaView.is_daemon` which enables or disables if a view should be a daemon or
get removed from the server.

Because this potentially changes the flow of user application code, the new
behavior is only active when `LonaView.STOP_DAEMON_WHEN_VIEW_FINISHES` is set
to `False`.

Signed-off-by: Florian Scherf <mail@florianscherf.de>