tests: rendering: fix start and stop button
Previously, both the start and the stop button got disabled when the mode was
set to `Manual`. This is not correct, because the start button should be
clickable only if the mode is set to `Auto` and the test loop is not running.
The stop button should be clickable, only if the mode is set to `Auto` and the
test loop is running.

Signed-off-by: Florian Scherf <mail@florianscherf.de>