tests: rendering: styling: fix CSS property name issues
Previously, the styling tests used the CSS properties `display`, `position` and
`z-index`. To check the actual CSS values in the browser, `getComputedStyle()`
in JavaScript gets used.

In Chrome, `getComputedStyle()` strips dashes from CSS property names, and
formats them to camel-case (`z-index` -> `zIndex`). This makes the test crash
in all browsers but Chrome.

To fix that, and avoid browser-specific code, this patch changes the test to
use `top`, 'right`, `bottom` and `left` for styling tests.

Signed-off-by: Florian Scherf <mail@florianscherf.de>