tests: rendering: fix JavaScript snippet to retrieve computed style
Previously, `getComputedStyle()` got used plain, to retrieve computed CSS
styles from the browser. This does only work in Chrome, because Chrome treats
the return type of `getComputedStyle()` like a Map. All other browsers treat it
like an Array, which, when serialized to JSON, results in a Map, with the Array
indices as keys, instead of the CSS property names.

This patch adds a snippet, that converts the return type to an JSON
serializeable object.

Signed-off-by: Florian Scherf <mail@florianscherf.de>