client2: initial commit
This patch is the initial commit for the new client implementation for Lona2.

Client2 is a completely separate code base, to be fully backwards-compatible,
until Lona2 gets released. As for this patch, it is a one-to-one copy of the
old code though.

The new client can be enabled, by setting the feature flag `CLIENT_VERSION` to
`2` in the settings (default is `1`).
On startup the server calls a new added method `lona.compat.set_client_version`
which sets the configured version in the environment variable
`LONA_CLIENT_VERSION`.
This extra step through the environment is necessary, to communicate to
`lona.html` types, which client version is running, since they have no
access to the settings.

This patch updates the rendering test, by adding a decorator `client_version`
to indicate which rendering step is necessary for which client, and a new
method `lona.compat.get_client_version()` to check which client set.

Signed-off-by: Florian Scherf <mail@florianscherf.de>