collect static: fix handling of overlapping directories
Previously collect-static crashed, when two static directories contained the
same sub directory.

  Example:
      project/static-dir-1/directory/file.txt
      project/static-dir-2/directory/file.txt

On Python versions after 3.7, this was fixed by setting the `dirs_exist_ok`
flag, in the `shutil.copytree()` call in collect-static.

    (https://docs.python.org/3/library/shutil.html#shutil.copytree)

Because this flag does not exist on Python 3.7, this patch adds code that
emulates this feature, and a check which implementation should be used.

This patch also adds a test for this case.

Signed-off-by: Florian Scherf <mail@florianscherf.de>