client2: rendering: drop support for legacy widget API
Lona1 defines two types of nodes: Nodes that can be rendered in the
browser (text nodes and elements), and collections of nodes that can be
rendered (widgets). This distinction was made to make components (for example a
pop-up component) with multiple root nodes possible.

This feature is useful in some cases but brings much complexity and error
potential into the JavaScript client. Also the implications of enforcing every
component to have exactly one root node, are not big enough to justify this
level of complexity.

This patch removes all widget rendering code from client2, updates
`lona.html.HTML` to return a node instead of a widget when client2 gets used,
and updates the rendering tests, to not test the legacy widget API when client2
is running.

Signed-off-by: Florian Scherf <mail@florianscherf.de>