client: select: send index of selected options instead of string value
Previously, `Select.value` always returned a string representation of the
original value of the selected version. This is surprising, and annoying when
using integer values, to represent ids from a database, for example.

This patch changes both clients, to send the index of an selected option,
instead of their value on the client. Which option is selected, then gets
correlated on the server.

To remain backwards compatible, the value still gets converted to a string. So
this patch does not fix the initial problem, but does the necessary ground work
to fix it in `lona.html.Select2` (comming soon).

Signed-off-by: Florian Scherf <mail@florianscherf.de>