html: parsing: use value properties instead of setting them as attribute
`value`, most of the time is used in nodes like `Select` or `TextInput`, and
is implemented as a high-level property.

Previously `value` got treated as an attribute:

    node = Node()
    node.attributes['value'] = value

This patch changes the parsing code to treat `value` as key word argument of
the node class, so all high-level properties get used:

    node = Node(value=value)

If the node does not implement a high-level property for `value`, the node
base-class falls back to setting `value` as an attribute.

Signed-off-by: Florian Scherf <mail@florianscherf.de>