setup response classes
Previously, Lona views used special dictionaries as responses, instead of
proper response classes like any other Python web framework.

This is a design that was part of Lona since the very beginning. In the early
days of this framework, views were simple functions, that needed almost no
imports. The idea was to use a Python standard data structure, so no classes
have to be imported, and no response class names have to be remembered.

This was a horrible idea, and lead to horrible code, namely the
`ResponseParser` code. Because the dictionaries could contain any key, they
had to be parsed.

This patch adds classes for any type of response, Lona supports, and a
drop-in-replacement for the response parser code, that converts dictionaries
into responses, to be backwards compatible, until Lona 2.

Signed-off-by: Florian Scherf <mail@florianscherf.de>