test project: frontend: add more precision to uptime
Previously, a JavaScript timestamp got used to indicate the frontends uptime.
The smallest unit of these timestamps are seconds, which is not enough
precision for browser tests, because within browser tests, clicks and checks
can be separated by milliseconds.

This patch changes the precision to milliseconds, using Pythons datetime
implementation. With this change, redirects or frontend restarts can be
tested precisely.

Signed-off-by: Florian Scherf <mail@florianscherf.de>