packaging: setup pyproject.toml
This patch sets up a `pyproject.toml` to fix multiple problems, with the
current packaging, at once:

1. Deprecation warnings

The current setup, using a legacy `setup.py`, produces this warning since
pip 23.0:

    DEPRECATION: lona is being installed using the legacy 'setup.py install'
    method, because it does not have a 'pyproject.toml' and the 'wheel' package
    is not installed. pip 23.1 will enforce this behaviour change. A possible
    replacement is to enable the '--use-pep517' option. Discussion can be found
    at https://github.com/pypa/pip/issues/8559

2. Problems with package data

On some systems, package data like the JavaScript client, or templates, are
missing, when Lona is installed using the git URL:

    pip install git+https://github.com/lona-web-org/lona.git

This seems to be only reproducible on Windows.

3. `client2` was not packaged properly before

`client2` was not part of the `MANIFEST.in` before.

This patch gets rid of the old `setup.py`, and all related files like
`MANIFEST.in`, and replaces all configuration with a `pyproject.toml` file.

Signed-off-by: Florian Scherf <mail@florianscherf.de>