views: fix redirects from input event handlers
This patch fixes a regression, introduced in

    ef19a2e45b08 ("setup response classes")

In ef19a2e45b08 `lona.responses.AbstractResponse` was introduced, as a new
data-structure for view responses. Since input event handlers can return
redirects, and HTTP redirects, which are now represented as sub classes of
`lona.response.AbstractResponse`, the view runtime has to check for these types
instead of dicts, like before ef19a2e45b08.

This was not covered by the original patch, and the regression was not catched
by the test, because the test did not enough checks, if a redirect from a
input event handler really worked, without any crashes.

This patch fixes the handling of `lona.responses.AbstractResponse` return
values of input event handlers, and updates the test, to check if a redirect,
actually worked.

Signed-off-by: Florian Scherf <mail@florianscherf.de>