tests: fix timing problems in simple daemon test
This test crashed frequently in the past, but went back to green when rerun,
most of the time. The final check, if the view was stopped correctly, was
always the failing check.

This seems to be a timing issue, because the browser seems not to wait long
enough, before checking if the string `view stopped` appeared.

This patch adds an `lona.pytest.eventually` call, to "fix" all timing issues by
simply retrying the checks.
This is no permanent solution, but will work for now.

Signed-off-by: Florian Scherf <mail@florianscherf.de>