client: lona namespace: remove public field declarations
JavaScript public field declarations are not supported in Safari versions
lower than 14.1. This caused problems on iPhone 6 and iPad mini 2.

This patch replaces all public field declarations with simple, object scoped
variables, set from the constructor, and replaces the `Lona` class, used as
a global namespace, with an object of the class `LonaNamespace`.
The renaming was done to prevent naming collisions between the class, and the
global object.

Further readings:
    https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Classes/Public_class_fields#browser_compatibility

Signed-off-by: Florian Scherf <mail@florianscherf.de>