lint: disable doc directory
The linter complains about multiple lines in the examples that should not be
changed, even though the linter is technically right. For example it complains
about unused variables, but the examples are meant to demonstrate that a
function returns a particular variable.

This patch disables the whole directory for now.

Signed-off-by: Florian Scherf <mail@florianscherf.de>