client2: add new widget API
Previously, the Lona client- and widget-API were very Pythonic, and used only
basic JavaScript class features before. In Lona 2, methods like
`Widget.data_updated` will be renamed to `Widget.onDataUpdated`. To facilitate
the transition, the new names and API will be added to Lona 1 and coexist with
the old API, to ensure compatibility until Lona 2.
This is done by adding a new JavaScript class, that abstracts the new and the
old widget API.

This patch adds `Widget.onDataUpdated` as a replacement for
`Widget.data_updated`, `Widget.destroy` as a replacement for
`Widget.deconstruct`, and adds all arguments of `Widget.setup` to
`Widget.constructor`, to make a separate setup stage, besides the JavaScript
constructor, obsolete in Lona 2.

This patch also adds a wrapper class for widget data, to
make helper methods like `WidgetData.set` and `WidgetData.get` possible in the
future.

Signed-off-by: Florian Scherf <mail@florianscherf.de>