client2: fix resolving of relative URLs in redirects
Previously, the resolving of relative URLs was a custom implementation and had
multiple weird quirks, and behaved differently than the browsers
implementation. That was confusing, because redirects, issued by Lona,
sometimes resulted in slightly different URLs than their link-counterpart.

This patch updates the client to use the browsers implementation of URL
resolving.

Signed-off-by: Florian Scherf <mail@florianscherf.de>