html: NodeList: reset: remove duplicate patches
`NodeList.reset()` gets called with a list of nodes when the `Node.nodes`
property gets set. `NodeList.reset()` then clears its node list and creates
a `lona.protocol.OPERATION.RESET` patch.

Previously, the code falsely create a patch for every new node, that contained
all following nodes. That resulted in a list of patches that would override
each other on the client.

This was no problem in the past, because the client has no checks if a node
was already sent by the server, and because the patches overwrote each other,
the HTML end-result always was correct.

This patch changes the code to create only one reset patch, instead of multiple
ones.

Signed-off-by: Florian Scherf <mail@florianscherf.de>