client2: rendering engine: add _cache_node method
Previously, the rendering engine wrote the node cache directly, to add a newly
rendered node, without checking if a already cached node gets overridden by it.
This is dangerous and covers up potential bugs in the rendering code on the
client and in the patch creation code on the server.

This patch adds `_cache_node()` to the rendering engine, which checks if a
node, with the given node id, is already cached, and throws an exception if so.

Signed-off-by: Florian Scherf <mail@florianscherf.de>